requirements.txt
langgraph
langchain
langchain-core
langchain-openai
python-dotenv
requests
pymongo 
Pillow
google-generativeai